server/requirements.txt
lark
pydot
pillow
nbformat
ipython
numpy
aalpy
graphviz
networkx
matplotlib
pandas
fastapi
pydantic
redis
openai
langchain
langchain_openai
uvicorn